abstracted content generator
