+++ app/react/stories/Modal.stories.tsx
@@ -3,6 +3,7 @@ import { Meta, StoryObj } from '@storybook/react';
 import { Modal } from 'V2/Components/UI/Modal';
 import { Button } from 'V2/Components/UI/Button';
 import { InformationCircleIcon } from '@heroicons/react/24/outline';
+import { GeneratedContent } from './helpers/GeneratedContent';
 
 const meta: Meta<typeof Modal> = {
   title: 'Components/Modal',
@@ -17,32 +18,6 @@ const Primary: Story = {
       <Modal size={args.size}>{args.children}</Modal>
     </div>
   ),
-};
-
-const ModalContent = () => {
-  const contents: React.ReactNode[] = [];
-
-  for (let index = 1; index < 8; index += 1) {
-    contents.push(
-      <>
-        <h1 className="font-bold">Item {index}</h1>
-        <p className="mb-1">
-          Fusce id mi eu mauris bibendum dignissim nec in sem. Sed ultrices varius mauris quis
-          placerat. Donec imperdiet sodales diam sed imperdiet. Aenean a nisl venenatis lectus
-          mattis pellentesque. Duis fermentum ante a ultricies feugiat. Proin dapibus luctus purus
-          id viverra. Aenean a aliquet nibh. Aenean facilisis justo quis sem auctor, nec mollis
-          tortor placerat. Cras eget enim mollis, mollis risus gravida, pharetra risus. Mauris
-          dapibus malesuada mi, quis ornare felis imperdiet eget. Donec sed quam non dolor sodales
-          hendrerit. Aenean suscipit, velit sed laoreet cursus, ante odio tristique lectus, a porta
-          eros felis eu sem. Curabitur eu gravida dolor. Ut iaculis lacus vitae libero viverra
-          interdum. Phasellus ac est consectetur, malesuada nisl nec, blandit lorem.
-        </p>
-        <hr className="mb-2" />
-      </>
-    );
-  }
-
-  return <>{contents.map(content => content)}</>;
 };
 
 const Basic = {
@@ -108,7 +83,7 @@ const LargeContent = {
           <Modal.CloseButton />
         </Modal.Header>
         <Modal.Body>
-          <ModalContent />
+          <GeneratedContent />
         </Modal.Body>
         <Modal.Footer>
           <Button styling="light" className="grow">
