+++ app/react/stories/Sidepanel.stories.tsx
@@ -5,36 +5,11 @@ import { Provider } from 'react-redux';
 import { LEGACY_createStore as createStore } from 'V2/shared/testingHelpers';
 import { Sidepanel } from 'V2/Components/UI';
 import { SidePanelProps } from 'app/V2/Components/UI/Sidepanel';
+import { GeneratedContent } from './helpers/GeneratedContent';
 
 const meta: Meta<typeof Sidepanel> = {
   title: 'Components/Sidepanel',
   component: Sidepanel,
-};
-
-const SidepanelContent = () => {
-  const contents: React.ReactNode[] = [];
-
-  for (let index = 1; index < 8; index += 1) {
-    contents.push(
-      <>
-        <h1 className="font-bold">Item {index}</h1>
-        <p className="mb-1">
-          Fusce id mi eu mauris bibendum dignissim nec in sem. Sed ultrices varius mauris quis
-          placerat. Donec imperdiet sodales diam sed imperdiet. Aenean a nisl venenatis lectus
-          mattis pellentesque. Duis fermentum ante a ultricies feugiat. Proin dapibus luctus purus
-          id viverra. Aenean a aliquet nibh. Aenean facilisis justo quis sem auctor, nec mollis
-          tortor placerat. Cras eget enim mollis, mollis risus gravida, pharetra risus. Mauris
-          dapibus malesuada mi, quis ornare felis imperdiet eget. Donec sed quam non dolor sodales
-          hendrerit. Aenean suscipit, velit sed laoreet cursus, ante odio tristique lectus, a porta
-          eros felis eu sem. Curabitur eu gravida dolor. Ut iaculis lacus vitae libero viverra
-          interdum. Phasellus ac est consectetur, malesuada nisl nec, blandit lorem.
-        </p>
-        <hr className="mb-2" />
-      </>
-    );
-  }
-
-  return <>{contents.map(content => content)}</>;
 };
 
 const SidePanelContainer = (args: SidePanelProps) => {
@@ -88,7 +63,7 @@ const SidePanelContainer = (args: SidePanelProps) => {
 
             <button
               type="button"
-              className="p-1 text-white border-2 rounded border-primary-400 bg-primary-400"
+              className="p-1 text-white rounded border-2 border-primary-400 bg-primary-400"
               onClick={() => setShowSidepanel(!showSidepanel)}
             >
               Open/Close sidepanel
@@ -101,7 +76,7 @@ const SidePanelContainer = (args: SidePanelProps) => {
             closeSidepanelFunction={() => setShowSidepanel(false)}
             size={args.size}
           >
-            <SidepanelContent />
+            <GeneratedContent />
           </Sidepanel>
         </div>
       </div>
